Fix up more GCC 8.2.1 warnings
Also add a network related KConfig option and adjust the minimum sizes
of some pools to meet the framework's own requirements.

Resolves: LE-12588

+++ framework/liblegato/cdata.c
@@ -29,7 +29,7 @@ unsigned int le_cdata_GetInstance
 
      for (i = 0; cdataPtr[i].key >= 0; ++i)
      {
-         if (cdataPtr[i].key == componentKey)
+         if ((unsigned int) cdataPtr[i].key == componentKey)
          {
              return cdataPtr[i].instance;
          }
